♻️refactor : 중복 코드 모듈 분리
