+++ js/validation.js
@@ -1,0 +1,40 @@
+export function isValidEmail(email) {
+  return /^([a-zA-Z0-9_\.-]+)@([a-zA-Z0-9_\-]+)\.([a-zA-Z]{2,})$/.test(email);
+}
+
+export function createErrorMessage(message, inputElement) {
+  const newError = document.createElement("p");
+  newError.textContent = message;
+  newError.classList.add("error-message");
+
+  inputElement.classList.add("error");
+  inputElement.parentElement.appendChild(newError);
+}
+
+export function validateEmail(emailInput) {
+  const emailError = emailInput.parentElement.querySelector(".error-message");
+
+  if (emailError) {
+    emailError.remove();
+    emailInput.classList.remove("error");
+  }
+
+  if (emailInput.value === "") {
+    createErrorMessage("이메일을 입력해 주세요.", emailInput);
+  } else if (!isValidEmail(emailInput.value)) {
+    createErrorMessage("올바른 이메일 주소가 아닙니다.", emailInput);
+  }
+}
+
+export function validatePassword(passwordInput) {
+  const passwordError = passwordInput.parentElement.querySelector(".error-message");
+
+  if (passwordError) {
+    passwordError.remove();
+    passwordInput.classList.remove("error");
+  }
+
+  if (passwordInput.value === "") {
+    createErrorMessage("비밀번호를 입력해 주세요.", passwordInput);
+  }
+}
