+++ signin.html
@@ -42,6 +42,6 @@
       </div>
     </div>
 
-    <script src="./js/sign.js"></script>
+    <script type="module" src="./js/signin.js"></script>
   </body>
 </html>
